+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Features
 
-* enhance UI components and improve responsiveness across various pages ([dc45aef](https://github.com/DiluDevX/deliveroo-clone-web/commit/dc45aef6f025877e1889519deaadcd8fb49ed6e4))
+- enhance UI components and improve responsiveness across various pages ([dc45aef](https://github.com/DiluDevX/deliveroo-clone-web/commit/dc45aef6f025877e1889519deaadcd8fb49ed6e4))
 
 ## [1.4.1](https://github.com/DiluDevX/deliveroo-clone-web/compare/v1.4.0...v1.4.1) (2026-05-28)